requirements.txt
aniso8601==7.0.0
asgiref==3.5.2
Django==4.1.1
graphene
graphene-django
graphql-core==2.3.2
graphql-relay
promise==2.3
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
django-filter
tzdata==2022.2
djangorestframework



